requirements.txt
transformers==4.43.2
accelerate==0.30.1
datasets==2.19.1
sentencepiece==0.2.0

numpy
psutil
tqdm

llama-cpp-python==0.2.84
